Convert VersionCompat into a library, and introduce a CompareVersions tool
This makes it easier to compare versions. Sample command lines:

--file1=path/to/old.dll --file2=path/to/new.dll
--package=Google.Cloud.Speech.V1 --version1=latest --version2=local
--package=Google.Cloud.Speech.V1 --version1=1.1.0 --version2=1.2.0 --framework=netstandard1.5
