+++ tools/Google.Cloud.Tools.CompareVersions/OutputBehavior.cs
@@ -1,0 +1,29 @@
+﻿// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
+
+namespace Google.Cloud.Tools.CompareVersions
+{
+    [Flags]
+    internal enum OutputBehavior
+    {
+        None = 0,
+        Major = 1,
+        Minor = 2,
+
+        // To allow all combinations to be set with a single flag value
+        MajorAndMinor = 3,
+    }
+}
