+++ tools/Google.Cloud.Tools.VersionCompat/Google.Cloud.Tools.VersionCompat.csproj
@@ -1,16 +1,14 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netstandard2.0</TargetFramework>
     <LangVersion>latest</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="CommandLineParser" Version="2.6.0" />
     <PackageReference Include="Mono.Cecil" Version="0.10.3" />
-    <PackageReference Include="NuGet.Client" Version="4.2.0" />
     <PackageReference Include="SharpCompress" Version="0.24.0" />
+    <PackageReference Include="System.Collections.Immutable" Version="1.7.0" />
   </ItemGroup>
 
 </Project>
